cmd/geth: set and get managers and rootchains from file server
